+++ sim.c
@@ -1,12 +1,12 @@
 #include "sim.h"
 
 #define C_IMPL
-extern void asm_do_row(Cell *, Cell *, uint32_t, uint32_t, uint32_t);
+extern void asm_do_row(Cell *, Cell *, size_t, size_t, size_t);
 
 /**
  * gets x mod N (works for negative numbers as well! Use this instead of %)
  */
-uint32_t mod(int x, uint32_t N)
+size_t mod(int x, size_t N)
 {
     return (x + x / N * N) % N;
 }
@@ -14,7 +14,7 @@ uint32_t mod(int x, uint32_t N)
 /**
  * process one row of the board
  */
-static void do_row(Cell *dest, Cell *src, uint32_t row, uint32_t rows, uint32_t cols)
+static void do_row(Cell *dest, Cell *src, size_t row, size_t rows, size_t cols)
 {
     // TODO
 }
